fix: Line is no longer keeping appropriate size on replay-drawing
